backend/requirements.txt
opik
fastapi
uvicorn
pydantic
openai
python-dotenv
